fix(inventory-advance-search-field): fix changing advance search won't apply
